* Recycler view order reverted

* Recycler view order reverted

* Recycler view order reverted

* update

* indented

* Recycler view order reverted

* Recycler view order reverted

* update

* indented

* Corrected your activity mistake

* Recycler view order reverted

* update

* indented

* update

* indented

* Corrected your activity mistake

* Corrected your activity mistake

* Changed image delete on post delete

* Changed image delete on post delete

* pull (#1)

* changes

* changes

* changes

* changes

* Reversed Recycler and added PostId's

* Commit

* commit

* double card layout

* double card layout

* commit

* merge

* main activity chagne

* main activity chagne

* my changes

* my changes

* my changes

* Major changes

* Pull please

* Pull please

+++ app/src/main/java/com/example/abhinav/buzzer/HomeSingleActivity.java
@@ -17,6 +17,8 @@ import com.google.firebase.database.DataSnapshot;
 import com.google.firebase.database.DatabaseError;
 import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
+import com.google.firebase.database.MutableData;
+import com.google.firebase.database.Transaction;
 import com.google.firebase.database.ValueEventListener;
 import com.google.firebase.storage.FirebaseStorage;
 import com.google.firebase.storage.StorageReference;
@@ -93,24 +95,36 @@ public class HomeSingleActivity extends AppCompatActivity {
         mHomeSingleRemoveBtn.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-
-                mDatabase.child(mPost_key).removeValue();
-               // if(hasImage==1){
-                mStorage= FirebaseStorage.getInstance().getReference().child("Posts/"+postId);
-                mStorage.delete().addOnSuccessListener(new OnSuccessListener<Void>() {
+                mDatabase.addListenerForSingleValueEvent(new ValueEventListener() {
                     @Override
-                    public void onSuccess(Void aVoid) {
-                        Toast.makeText(HomeSingleActivity.this,"Removed Succesfully",Toast.LENGTH_SHORT).show();                    }
+                    public void onDataChange(DataSnapshot dataSnapshot) {
+                      Long val=(Long)dataSnapshot.child(mPost_key).child("With_image").getValue();
+                        if(val==1)
+                        {
+                            mStorage= FirebaseStorage.getInstance().getReference().child("Posts/"+postId);
+                            mStorage.delete().addOnSuccessListener(new OnSuccessListener<Void>() {
+                                @Override
+                                public void onSuccess(Void aVoid) {
+                                    Toast.makeText(HomeSingleActivity.this,"Removed Succesfully",Toast.LENGTH_SHORT).show();                    }
 
-                }).addOnFailureListener(new OnFailureListener() {
+                            }).addOnFailureListener(new OnFailureListener() {
+                                @Override
+                                public void onFailure(@NonNull Exception exception) {
+                                    Toast.makeText(HomeSingleActivity.this,"Failed",Toast.LENGTH_SHORT).show();
+                                }
+                            });
+                        }
+                        else
+                            Toast.makeText(HomeSingleActivity.this,"Removed Successfully",Toast.LENGTH_SHORT).show();
+                    }
+
                     @Override
-                    public void onFailure(@NonNull Exception exception) {
-                        Toast.makeText(HomeSingleActivity.this,"Failed",Toast.LENGTH_SHORT).show();
+                    public void onCancelled(DatabaseError databaseError) {
+
                     }
                 });
-                //}
-               // else
-              //      Toast.makeText(HomeSingleActivity.this,"Removed Succesfully",Toast.LENGTH_SHORT).show();
+
+                mDatabase.child(mPost_key).removeValue();
 
                 Intent mainIntent = new Intent(HomeSingleActivity.this, MainActivity.class);
                 startActivity(mainIntent);
